Add logger implementations and interface
