+++ dotnet/logger/IAuthenticationLogger.cs
@@ -1,0 +1,13 @@
+namespace Microsoft.Commerce.Payments.PXCommon
+{
+    using System;
+
+    public interface IAuthenticationLogger
+    {
+        void LogMiseTokenValidationResult(
+            MiseTokenValidationResult result,
+            long latency,
+            Exception? exception,
+            string incomingRequestId);
+    }
+}
